requirements.txt
streamlit
Pillow
numpy
scipy
scikit-image
pandas
transformers
torch
ultralytics
requests
+easyocr
